+++ src/nice_elapsed/mod.rs
@@ -4,7 +4,6 @@
 Note: this module is "in development". It is subject to change, and may eventually be spun off into its own crate.
 */
 
-use crate::macros;
 use std::{
 	fmt,
 	hash::{
@@ -23,7 +22,17 @@ use std::{
 /// # Array Size.
 const SIZE: usize = 39;
 
-
+/// # Helper: `AsRef` and `Borrow`.
+macro_rules! as_ref_borrow_cast {
+	($($cast:ident $ty:ty),+ $(,)?) => ($(
+		impl AsRef<$ty> for NiceElapsed {
+			fn as_ref(&self) -> &$ty { self.$cast() }
+		}
+		impl ::std::borrow::Borrow<$ty> for NiceElapsed {
+			fn borrow(&self) -> &$ty { self.$cast() }
+		}
+	)+);
+}
 
 /// # Helper: Generate Impl
 macro_rules! elapsed_from {
@@ -72,11 +81,7 @@ pub struct NiceElapsed {
 	len: usize,
 }
 
-macros::as_ref_borrow_cast!(
-	NiceElapsed:
-		as_bytes [u8],
-		as_str str,
-);
+as_ref_borrow_cast!(as_bytes [u8], as_str str);
 
 impl Default for NiceElapsed {
 	#[inline]
@@ -103,7 +108,12 @@ impl fmt::Debug for NiceElapsed {
 	}
 }
 
-macros::display_str!(as_str NiceElapsed);
+impl fmt::Display for NiceElapsed {
+	#[inline]
+	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+		f.write_str(self.as_str())
+	}
+}
 
 impl Eq for NiceElapsed {}