+++ src/nice_int/nice_u8.rs
@@ -2,16 +2,21 @@
 # Dactyl: Nice u8.
 */
 
+use crate::NiceWrapper;
 use std::num::NonZeroU8;
 
 
 
 /// # Total Buffer Size.
+///
+/// 255 needs no commas, so is only three bytes.
 const SIZE: usize = 3;
 
+/// # Zero.
+const ZERO: [u8; SIZE] = [b'0', b'0', b'0'];
 
 
-#[derive(Debug, Clone, Copy)]
+
 /// `NiceU8` provides a quick way to convert a `u8` into a formatted byte
 /// string for e.g. printing.
 ///
@@ -26,16 +31,40 @@ const SIZE: usize = 3;
 ///     "231"
 /// );
 /// ```
-pub struct NiceU8 {
-	inner: [u8; SIZE],
-	from: usize,
-}
+///
+/// ## Traits
+///
+/// Rustdoc doesn't do a good job at documenting type alias implementations, but
+/// `NiceU8` has a bunch, including:
+///
+/// * `AsRef<[u8]>`
+/// * `AsRef<str>`
+/// * `Borrow<[u8]>`
+/// * `Borrow<str>`
+/// * `Clone`
+/// * `Copy`
+/// * `Default`
+/// * `Deref<Target=[u8]>`
+/// * `Display`
+/// * `Eq` / `PartialEq`
+/// * `Hash`
+/// * `Ord` / `PartialOrd`
+///
+/// You can instantiate a `NiceU8` with:
+///
+/// * `From<u8>`
+/// * `From<Option<u8>>`
+/// * `From<NonZeroU8>`
+/// * `From<Option<NonZeroU8>>`
+///
+/// When converting from a `None`, the result will be equivalent to zero.
+pub type NiceU8 = NiceWrapper<SIZE>;
 
 impl Default for NiceU8 {
 	#[inline]
 	fn default() -> Self {
 		Self {
-			inner: [b'0', b'0', b'0'],
+			inner: ZERO,
 			from: SIZE,
 		}
 	}
@@ -46,7 +75,7 @@ impl From<u8> for NiceU8 {
 	#[allow(unsafe_code)]
 	fn from(num: u8) -> Self {
 		if 99 < num {
-			let mut inner = [b'0', b'0', b'0'];
+			let mut inner = ZERO;
 			unsafe { super::write_u8_3(inner.as_mut_ptr(), num as u16); }
 			Self {
 				inner,
@@ -54,7 +83,7 @@ impl From<u8> for NiceU8 {
 			}
 		}
 		else if 9 < num {
-			let mut inner = [b'0', b'0', b'0'];
+			let mut inner = ZERO;
 			unsafe {
 				std::ptr::copy_nonoverlapping(
 					crate::double(num as usize),
@@ -76,9 +105,8 @@ impl From<u8> for NiceU8 {
 	}
 }
 
-// A few Macro traits.
-super::impl_nice_nonzero_int!(NiceU8: NonZeroU8);
-super::impl_nice_int!(NiceU8);
+super::nice_from_nz!(NiceU8, NonZeroU8);
+super::nice_from_opt!(NiceU8);
 
 impl NiceU8 {
 	#[must_use]
@@ -88,7 +116,7 @@ impl NiceU8 {
 	/// This is equivalent to zero.
 	pub const fn min() -> Self {
 		Self {
-			inner: [b'0', b'0', b'0'],
+			inner: ZERO,
 			from: SIZE - 1,
 		}
 	}
@@ -192,6 +220,12 @@ mod tests {
 		assert_eq!(NiceU8::default().as_bytes(), <&[u8]>::default());
 		assert_eq!(NiceU8::default().as_str(), "");
 
+		// Test some Option variants.
+		let foo: Option<u8> = None;
+		assert_eq!(NiceU8::min(), NiceU8::from(foo));
+		let foo = Some(13_u8);
+		assert_eq!(NiceU8::from(13_u8), NiceU8::from(foo));
+
 		// Check ordering too.
 		let one = NiceU8::from(10);
 		let two = NiceU8::from(90);
@@ -205,6 +239,12 @@ mod tests {
 		assert_eq!(NiceU8::min(), NiceU8::from(NonZeroU8::new(0)));
 		assert_eq!(NiceU8::from(50_u8), NiceU8::from(NonZeroU8::new(50)));
 		assert_eq!(NiceU8::from(50_u8), NiceU8::from(NonZeroU8::new(50).unwrap()));
+
+		// Test some Option variants.
+		let foo: Option<NonZeroU8> = None;
+		assert_eq!(NiceU8::from(foo), NiceU8::min());
+		let foo = NonZeroU8::new(13);
+		assert_eq!(NiceU8::from(13_u8), NiceU8::from(foo));
 	}
 
 	#[test]
@@ -238,7 +278,7 @@ mod tests {
 	#[test]
 	fn t_as() {
 		let num = NiceU8::from(253);
-		assert_eq!(num.as_str(), num.as_string());
-		assert_eq!(num.as_bytes(), num.as_vec());
+		assert_eq!(num.as_str(), String::from(num));
+		assert_eq!(num.as_bytes(), Vec::<u8>::from(num));
 	}
 }